03_BilderLabling/requirements.txt
numpy~=1.25.1
matplotlib~=3.7.2
tqdm~=4.65.0
ipython~=8.14.0
jupyter~=1.0.0
ipywidgets~=8.1.0
pyside6~=6.5.2
Pillow~=10.0.1
